Add -Werror flag to SRC_SOURCES and fix compilation warnings
- Configure CMakeLists.txt to treat warnings as errors for SRC_SOURCES
- Remove unused variable num_calls in decode_ft8.c
- Fix signed/unsigned comparison in gen_ft8.c by casting strlen() to int
- Increase log buffer size in main.c to prevent format truncation

+++ Src/gen_ft8.c
@@ -154,7 +154,7 @@ void clear_xmit_messages(void)
 static void set_text(char *text, const char *source, int field_id)
 {
 	strcpy(text, source);
-	for (int i = 0; i < strlen(text); ++i)
+	for (int i = 0; i < (int)strlen(text); ++i)
 	{
 		if (!isprint((int)text[i]))
 		{
